Steering AI system Working!!!
The arrive behavior is still bugged since jittering still happens on arrival, the states keep flipping rapidly between idle and dodge, and the sprite rotation is still not working properly.

However, The steering system and all the associated math are up and running. Next thing to work on is the wall collision or path following stuff.